Move LLMType to types/llm.ts
